requirements.txt
alembic==1.4.3
amqp==5.0.2
billiard==3.6.3.0
cachelib==0.1.1
celery==5.0.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cryptography==3.3.1
dnspython==1.16.0
eventlet==0.29.1
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
greenlet==0.4.17
idna==2.10
importlib-metadata==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==8.0.1
prompt-toolkit==3.0.8
pycparser==2.20
pycryptodomex==3.9.4
PyMySQL==0.10.1
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
qiniu==7.3.0
redis==3.5.3
requests==2.25.0
ronglian-sms-sdk==1.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
